GrafanaAPIInterface/requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-WTF==0.15.1
idna==2.10
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
